Refactor cache busting
This leverages the `last_changed` cache key from WP core, which is
changed when posts are created, updated, or deleted.

The value from the that cache key is stored along with the query
results, and is checked against the current value to determine
whether the cached query results should be returned or reset.